# Support attrs decorators even if they are imported from attrs