--- a/astroid/brain/brain_attrs.py
+++ b/astroid/brain/brain_attrs.py
@@ -8,15 +8,9 @@ Astroid hook for the attrs library
 Without this hook pylint reports unsupported-assignment-operation
 for attrs classes
 """
+from astroid.helpers import safe_infer
 from astroid.manager import AstroidManager
-from astroid.nodes.node_classes import (
-    AnnAssign,
-    Assign,
-    AssignName,
-    Call,
-    ImportFrom,
-    Unknown,
-)
+from astroid.nodes.node_classes import AnnAssign, Assign, AssignName, Call, Unknown
 from astroid.nodes.scoped_nodes import ClassDef
 
 ATTRIB_NAMES = frozenset(
@@ -47,18 +41,10 @@ def is_decorated_with_attrs(node, decorator_names=ATTRS_NAMES) -> bool:
             decorator_attribute = decorator_attribute.func
         if decorator_attribute.as_string() in decorator_names:
             return True
-        # Check if by including the origin module of the decorator -
-        # and concatenating them we get one of ATTRS_NAMES
-        decorator_name = decorator_attribute.as_string()
-        local_value = decorator_attribute.scope().locals.get(decorator_name)
 
-        # TODO: hack, as sometimes its a list?
-        if isinstance(local_value, ImportFrom) or (
-            isinstance(local_value, list) and isinstance(local_value[0], ImportFrom)
-        ):
-            decorator_module = local_value[0].modname
-            if f"{decorator_module}.{decorator_name}" in ATTRS_NAMES:
-                return True
+        inferred = safe_infer(decorator_attribute)
+        if inferred and inferred.root().name == "attr._next_gen":
+            return True
     return False