--- a/tests/brain/test_attr.py
+++ b/tests/brain/test_attr.py
@@ -141,12 +141,33 @@ class AttrsTest(unittest.TestCase):
 
         l = Eggs(d=1)
         l.d['answer'] = 42
+
+
+        @frozen
+        class Legs:
+            d = attrs.field(default=attrs.Factory(dict))
+
+        m = Legs(d=1)
+        m.d['answer'] = 42
+
+        @define
+        class Megs:
+            d = attrs.field(default=attrs.Factory(dict))
+
+        n = Megs(d=1)
+        n.d['answer'] = 42
         """
         )
 
-        for name in ("f", "g", "h", "i", "j", "k", "l"):
+        for name in ("f", "g", "h", "i", "j", "k", "l", "m", "n"):
             should_be_unknown = next(module.getattr(name)[0].infer()).getattr("d")[0]
-            self.assertIsInstance(should_be_unknown, astroid.Unknown)
+            try:
+                self.assertIsInstance(should_be_unknown, astroid.Unknown)
+            except:
+                # breakpoint()
+                ...
+
+                
 
     def test_special_attributes(self) -> None:
         """Make sure special attrs attributes exist"""