--- a/tests/brain/test_attr.py
+++ b/tests/brain/test_attr.py
@@ -91,6 +91,7 @@ class AttrsTest(unittest.TestCase):
             """
         import attrs
         from attrs import field, mutable, frozen, define
+        from attrs import mutable as my_mutable
 
         @attrs.define
         class Foo:
@@ -151,15 +152,29 @@ class AttrsTest(unittest.TestCase):
         m.d['answer'] = 42
 
         @define
-        class Megs:
+        class FooBar:
             d = attrs.field(default=attrs.Factory(dict))
 
-        n = Megs(d=1)
+        n = FooBar(d=1)
         n.d['answer'] = 42
+
+        @mutable
+        class BarFoo:
+            d = attrs.field(default=attrs.Factory(dict))
+
+        o = BarFoo(d=1)
+        o.d['answer'] = 42
+
+        @my_mutable
+        class FooFoo:
+            d = attrs.field(default=attrs.Factory(dict))
+
+        p = FooFoo(d=1)
+        p.d['answer'] = 42
         """
         )
 
-        for name in ("f", "g", "h", "i", "j", "k", "l", "m", "n"):
+        for name in ("f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p"):
             should_be_unknown = next(module.getattr(name)[0].infer()).getattr("d")[0]
             self.assertIsInstance(should_be_unknown, astroid.Unknown)