This PR is a POC for improving the way that astroid detects `attrs` decorators.
I am assuming that adding `frozen`, `define`, and `mutable` to `ATTRS_NAMES` was considered and was decided against - otherwise, it should probably be preferred

## Type of Changes

|     | Type                   |
| --- | ---------------------- |
| ✓   | :bug: Bug fix          |

## Description

Closes https://github.com/PyCQA/pylint/issues/7884