@DanielNoord 2023-03-29
@alonme Sorry for pushing to your branch but I think this might be a better solution. Instead of using some hack we actually use our inference brain.

@Pierre-Sassoulas Do we have a good way to measure the performance impact of this change?