@Pierre-Sassoulas 2023-03-29
No we'd need a performance primer, this has been on my mind for quite some time now but there is always something else to do :)