@alonme 2023-03-30
@DanielNoord  - No worries, this is of course much better.

Tell me if i can help with anything else