@DanielNoord 2023-03-30
> @DanielNoord - No worries, this is of course much better.
> 
> Tell me if i can help with anything else

Well, I'm not sure if you have a large codebase at work or for personal projects but it would be really good to see how long `pylint` takes with this PR merged into `astroid` for a project that currently takes comparatively long to lint already. 

Oh, and we need a changelog entry :smile: