@Pierre-Sassoulas 2023-03-30
We do have a benchmark test in pylint, it's not automated but you can check the difference between two versions of editable install of astroid. clone pylint then pip install -e your local astroid, launch ``pytest -k benchmark`` once for main astroid, once for this branch and compare the result.