@alonme 2023-03-30
Original

```
---------------------------------------------------------------------------------------------------------------- benchmark 'baseline': 11 tests ----------------------------------------------------------------------------------------------------------------
Name (time in us)                                                  Min                       Max                      Mean                 StdDev                    Median                    IQR            Outliers         OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_baseline_benchmark_j1                                    642.9770 (1.0)          1,428.7080 (1.0)            739.2487 (1.0)         112.0223 (1.0)            694.5560 (1.0)         101.8795 (1.0)         89;53  1,352.7248 (1.0)         613           1
test_baseline_benchmark_j1_all_checks_single_file         116,752.9580 (181.58)     211,995.9280 (148.38)     141,993.9449 (192.08)   29,315.4609 (261.69)     136,810.1150 (196.97)   11,071.6855 (108.67)        1;1      7.0426 (0.01)          8           1
test_baseline_lots_of_files_j1_empty_checker              223,609.7690 (347.77)     232,891.4640 (163.01)     228,809.2736 (309.52)    3,556.4965 (31.75)      228,393.3320 (328.83)    4,818.5200 (47.30)         2;0      4.3705 (0.00)          5           1
test_baseline_lots_of_files_j1                            225,234.3440 (350.30)     280,985.0320 (196.67)     241,983.4922 (327.34)   22,436.6740 (200.29)     234,849.1460 (338.13)   21,192.6762 (208.02)        1;0      4.1325 (0.00)          5           1
test_baseline_benchmark_j1_all_checks_lots_of_files       242,627.6570 (377.35)     261,145.6920 (182.78)     252,308.4546 (341.30)    7,917.0373 (70.67)      249,653.8560 (359.44)   13,229.2232 (129.85)        2;0      3.9634 (0.00)          5           1
test_baseline_benchmark_j1_single_working_checker         504,609.7670 (784.80)     508,595.6660 (355.98)     506,773.7986 (685.53)    1,528.7661 (13.65)      506,435.9390 (729.15)    2,057.6145 (20.20)         2;0      1.9733 (0.00)          5           1
test_baseline_benchmark_check_parallel_j2                 860,153.1770 (>1000.0)    873,649.6280 (611.50)     867,640.9618 (>1000.0)   5,228.4055 (46.67)      868,421.8980 (>1000.0)   7,659.0113 (75.18)         2;0      1.1526 (0.00)          5           1
test_baseline_benchmark_j2                                861,621.1170 (>1000.0)    886,040.1400 (620.17)     869,244.8098 (>1000.0)   9,799.4752 (87.48)      866,038.2450 (>1000.0)  10,326.3895 (101.36)        1;0      1.1504 (0.00)          5           1
test_baseline_lots_of_files_j2                          1,095,141.9380 (>1000.0)  1,109,569.9160 (776.62)   1,101,237.5432 (>1000.0)   5,580.7763 (49.82)    1,099,350.8350 (>1000.0)   7,628.3482 (74.88)         2;0      0.9081 (0.00)          5           1
test_baseline_lots_of_files_j2_empty_checker            1,153,549.8460 (>1000.0)  1,235,064.6380 (864.46)   1,195,213.4776 (>1000.0)  34,162.0979 (304.96)   1,201,637.4480 (>1000.0)  58,508.2465 (574.29)        2;0      0.8367 (0.00)          5           1
test_baseline_benchmark_j2_single_working_checker       1,396,959.8420 (>1000.0)  1,442,936.7230 (>1000.0)  1,412,449.0406 (>1000.0)  18,095.8019 (161.54)   1,405,713.4690 (>1000.0)  19,772.2875 (194.08)        1;0      0.7080 (0.00)          5           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

Legend:
  Outliers: 1 Standard Deviation from Mean; 1.5 IQR (InterQuartile Range) from 1st Quartile and 3rd Quartile.
  OPS: Operations Per Second, computed as 1 / Mean
```

New Astroid
```
---------------------------------------------------------------------------------------------------------------- benchmark 'baseline': 11 tests ----------------------------------------------------------------------------------------------------------------
Name (time in us)                                                  Min                       Max                      Mean                 StdDev                    Median                    IQR            Outliers         OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_baseline_benchmark_j1                                    641.5790 (1.0)          1,328.2370 (1.0)            736.4457 (1.0)         106.5480 (1.0)            691.5840 (1.0)          95.7635 (1.0)         69;49  1,357.8734 (1.0)         473           1
test_baseline_benchmark_j1_all_checks_single_file         111,243.2120 (173.39)     189,752.4330 (142.86)     122,381.1555 (166.18)   27,267.7914 (255.92)     112,297.6815 (162.38)    3,274.8270 (34.20)         1;1      8.1712 (0.01)          8           1
test_baseline_lots_of_files_j1_empty_checker              224,901.3490 (350.54)     246,673.0210 (185.71)     235,200.4922 (319.37)   10,376.5844 (97.39)      234,226.3730 (338.68)   20,141.1570 (210.32)        1;0      4.2517 (0.00)          5           1
test_baseline_lots_of_files_j1                            226,492.3970 (353.02)     233,275.2050 (175.63)     228,564.7700 (310.36)    2,752.9989 (25.84)      228,071.7350 (329.78)    2,900.1143 (30.28)         1;0      4.3751 (0.00)          5           1
test_baseline_benchmark_j1_all_checks_lots_of_files       231,990.3950 (361.59)     241,504.4150 (181.82)     236,295.9408 (320.86)    4,392.1657 (41.22)      235,916.5710 (341.12)    8,304.6982 (86.72)         1;0      4.2320 (0.00)          5           1
test_baseline_benchmark_j1_single_working_checker         505,705.3120 (788.22)     510,921.8300 (384.66)     509,039.6542 (691.21)    2,207.7081 (20.72)      510,081.9000 (737.56)    3,334.7282 (34.82)         1;0      1.9645 (0.00)          5           1
test_baseline_benchmark_check_parallel_j2                 856,387.9930 (>1000.0)    880,670.6260 (663.04)     866,981.7080 (>1000.0)   8,974.2807 (84.23)      865,482.9950 (>1000.0)  10,696.0433 (111.69)        2;0      1.1534 (0.00)          5           1
test_baseline_benchmark_j2                                857,057.6400 (>1000.0)    916,018.9590 (689.65)     873,404.7904 (>1000.0)  24,171.7238 (226.86)     865,202.7050 (>1000.0)  19,876.2832 (207.56)        1;1      1.1449 (0.00)          5           1
test_baseline_lots_of_files_j2                          1,079,281.9450 (>1000.0)  1,105,041.7630 (831.96)   1,094,832.0024 (>1000.0)   9,925.2562 (93.15)    1,098,884.7370 (>1000.0)  12,321.1958 (128.66)        2;0      0.9134 (0.00)          5           1
test_baseline_lots_of_files_j2_empty_checker            1,102,761.4370 (>1000.0)  1,121,359.2350 (844.25)   1,110,383.5888 (>1000.0)   8,500.3720 (79.78)    1,107,454.7770 (>1000.0)  15,350.8312 (160.30)        1;0      0.9006 (0.00)          5           1
test_baseline_benchmark_j2_single_working_checker       1,391,254.1400 (>1000.0)  1,403,989.3820 (>1000.0)  1,396,174.4972 (>1000.0)   5,111.8423 (47.98)    1,396,403.4950 (>1000.0)   7,196.2392 (75.15)         1;0      0.7162 (0.00)          5           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

Legend:
  Outliers: 1 Standard Deviation from Mean; 1.5 IQR (InterQuartile Range) from 1st Quartile and 3rd Quartile.
  OPS: Operations Per Second, computed as 1 / Mean
```